feat,refactor<BE> : 리그 Api 요청 방식 추가 및 게시글 내림차순 정렬 추가, 비밀번호 변경 로직 구현, 카카오 회원가입 로그인 구현

See merge request s07-webmobile1-sub2/S07P12B204!91

+++ backend/src/main/java/com/ssafy/cheertogether/article/dto/ArticleModifyRequest.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 public class ArticleModifyRequest {
 	private String title;
 	private String content;
-	private int leagueApiId;
+	private Long leagueApiId;
 }
